+++ src/helix/users/mod.rs
@@ -39,6 +39,6 @@ pub use get_user_block_list::{GetUserBlockListRequest, UserBlock};
 #[doc(inline)]
 pub use get_users::{GetUsersRequest, User};
 #[doc(inline)]
-pub use get_users_follows::{GetUsersFollowsRequest, UsersFollows};
+pub use get_users_follows::{FollowRelationship, GetUsersFollowsRequest, UsersFollows};
 #[doc(inline)]
 pub use unblock_user::{UnblockUser, UnblockUserRequest};